Add files via upload
This zip file contains the notebook(ImageAugmentation.ipynb) that has the function(dataAugmentation) the project used for data augmentation (including rotating the images and change the brightness). The zip file also includes a sample pictures file called Sample_imgs,  which is used in notebook to demonstrate the use of the function.